fix: pin ruff, and clear the lint CI had been failing on since v0.1.0
`ruff>=0.3` in [dev] meant CI resolved to whatever shipped that week
while contributors ran whatever they installed months ago. A linter's
default rule set moves between releases, so main went red on code nobody
had touched — 26 errors across files last edited in July, every push
since 2026-08-12, on a workflow whose failure had stopped meaning
anything.

Pinned to ruff==0.16.3, the version CI was already resolving to, and
cleared what it reports:

- 15 import blocks sorted (I001), the bulk of it
- 8 markup strings in list literals parenthesized (ISC004), so a missing
  comma can no longer hide as concatenation
- 2 subprocess.run calls say check=False (PLW1510), which is what both
  already meant: a non-zero exit is data here, not an exception
- one int(round(x)) that was already an int (RUF046), and re.S spelled
  re.DOTALL

Behaviour-neutral: 347 tests pass, and `ruff check src tests` — the
command CI runs — is clean. The four SIM115 in examples/ are left as
they are: CI does not lint them, and a one-line csv read is the point of
a teaching script.

+++ src/specthis/timefmt.py
@@ -11,7 +11,7 @@ from datetime import datetime, timezone
 
 def fmt_duration(seconds: float) -> str:
     """Wall time: ``4s``, ``3m 12s``, ``1h 04m``."""
-    s = int(round(seconds))
+    s = round(seconds)
     if s < 60:
         return f"{s}s"
     if s < 3600:
